% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADSIHT-package.R
\docType{package}
\name{ADSIHT-package}
\alias{ADSIHT-package}
\title{ADSIHT: Adaptive Double Sparse Iterative Hard Thresholding}
\description{
Solving high-dimensional double sparse linear regression via an iterative hard thresholding algorithm. Furthermore, the method is extended to jointly estimate multiple graphical models. For more details, please see \url{https://www.jmlr.org/papers/v25/23-0653.html} and \doi{10.48550/arXiv.2503.18722}.
}
\author{
\strong{Maintainer}: Yanhang Zhang \email{zhangyh98@ruc.edu.cn}

Authors:
\itemize{
  \item Zhifan Li \email{zhifanli@bimsa.cn}
  \item Jianxin Yin \email{jyin@ruc.edu.cn}
  \item Shixiang Liu \email{liushixiang_stat@ruc.edu.cn}
  \item Lingren Kong \email{lingrenkong@ruc.edu.cn}
}

}
\keyword{internal}
