% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADSIHT-package.R
\docType{package}
\name{ADSIHT-package}
\alias{ADSIHT-package}
\title{ADSIHT: Adaptive Double Sparse Iterative Hard Thresholding}
\description{
An efficient toolkit for solving the high-dimensinal double sparse linear regression via iterative hard threshlding algrithm. For more details, please see \url{https://arxiv.org/pdf/2305.04182}.
}
\author{
\strong{Maintainer}: Yanhang Zhang \email{zhangyh98@ruc.edu.cn}

Authors:
\itemize{
  \item Lingren Kong \email{lingrenkong@ruc.edu.cn}
}

}
\keyword{internal}
