% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADSIHT.R
\name{ADSIHT}
\alias{ADSIHT}
\title{Adaptive Double Sparse Iterative Hard Thresholding Algorithm (ADSIHT)}
\usage{
ADSIHT(
  x,
  y,
  group,
  s0,
  kappa = 0.9,
  ic.type = c("ebic", "sic", "bic", "aic", "loss"),
  ic.scale = 1,
  ic.coef = 1,
  L = 5,
  weight = rep(1, nrow(x))
)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{n \times p}; each row is an observation
vector and each column is a predictor.}

\item{y}{The response variable of \code{n} observations.}

\item{group}{A vector indicating which group each variable belongs to
For variables in the same group, they should be located in adjacent columns of \code{x}
and their corresponding index in \code{group} should be the same.
Denote the first group as \code{1}, the second \code{2}, etc.}

\item{s0}{A vector that controls the degrees with group.
Default is \code{d^((l-1)/(L-1)) : 1 \leq l \leq L}, where d is the maximum group size.}

\item{kappa}{A parameter that controls the rapid of the decrease of threshold. Default is 0.9.}

\item{ic.type}{The type of criterion for choosing the support size.
Available options are \code{"ebic"}, \code{"sic"}, \code{"bic"}, \code{"aic"} and \code{"loss"}.
Default is \code{"ebic"}.}

\item{ic.scale}{A non-negative value used for multiplying the penalty term
in information criterion. Default: \code{ic.scale = 1}.}
}
\value{

}
\description{
An implementation of the sparse group selection in linear regression model via ADSIHT.
}
\examples{

n <- 200
m <- 100
d <- 10
s <- 5
s0 <- 5
data <- gen.data(n, m, d, s, s0)
fit <- ADSIHT(data$x, data$y, data$group)
fit$A_out[which.min(fit$ic)]
}
\author{
Yanhang Zhang, Zhifan Li, Jianxin Yin.
}
